fix: IMA & DAI seek doesn't works after pre-roll Ad finished
Issue: engine switch to active after adBreakEnd but the lib load failed so should remove the listeners which cause the engine to return to works on lib failed loading.
Solution: change the active state to false after we finished everything and remove the listener which cause the decorator to keeps working.